sring-boot
final

+++ PI/src/main/java/tn/pi/spring/repository/PublicationRepository.java
@@ -1,19 +1,16 @@
 package tn.pi.spring.repository;
 
-import java.sql.Date;
-import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+
 //AUTHOR ABDESSALEM BENCHRIFA
 import org.springframework.stereotype.Repository;
 
-import tn.pi.spring.entity.Commentaire;
+
 import tn.pi.spring.entity.Publication;
 
 @Repository
-public interface PublicationRepository extends JpaRepository<Publication, Long> {
+public interface PublicationRepository extends JpaRepository<Publication, Integer> {
 	
 	
 	/*//search
